pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ops-center-cli"
version = "1.0.0"
description = "Command-line interface for Ops-Center infrastructure management"
readme = "README-CLI.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ops-Center Team", email = "ops-center@example.com"}
]
keywords = ["ops-center", "cli", "devops", "infrastructure", "management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
ops-center = "cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/ops-center"
Documentation = "https://ops-center.readthedocs.io"
Repository = "https://github.com/yourusername/ops-center"
"Bug Tracker" = "https://github.com/yourusername/ops-center/issues"

[tool.setuptools]
packages = ["cli", "cli.commands"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
